Update README.md
I've fiddled with markdown / HTML a bit hoping to find a way to get the images to align nicely with each other as well as have a caption under each of them, but I can only seem to do one simultaneously. So this would be a version where there are pictures for each of the authors (currently updated with placeholder pictures except for my own until I have everyones photos). I've also hosted the images on aws and I'll be able to do the same with others.